with customStringConfiguration
