+++ px-checkout/src/main/java/com/mercadopago/android/px/configuration/CustomStringConfiguration.java
@@ -1,15 +1,20 @@
 package com.mercadopago.android.px.configuration;
 
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.annotation.StringRes;
+import android.text.TextUtils;
+
 import com.mercadopago.android.px.R;
 
 public class CustomStringConfiguration {
 
     @StringRes private final int mainVerbStringResourceId;
+    @Nullable private final String customPaymentVaultTitle;
 
     /* default */ CustomStringConfiguration(@NonNull final Builder builder) {
         mainVerbStringResourceId = builder.mainVerbStringResourceId;
+        customPaymentVaultTitle = builder.customPaymentVaultTitle;
     }
 
     /**
@@ -22,8 +27,25 @@ public class CustomStringConfiguration {
         return mainVerbStringResourceId;
     }
 
+    /**
+     * Check if has
+     * @return true/false if customPaymentVaultTitle is (or not) empty
+     */
+    public boolean hasCustomPaymentVaultTitle() {
+        return !TextUtils.isEmpty(customPaymentVaultTitle);
+    }
+
+    /**
+     *
+     * @return Custom Payment Vault Title
+     */
+    public String getCustomPaymentVaultTitle() {
+        return customPaymentVaultTitle;
+    }
+
     public static class Builder {
         /* default */ int mainVerbStringResourceId;
+        /* default */ String customPaymentVaultTitle;
 
         public Builder() {
             mainVerbStringResourceId = R.string.px_main_verb;
@@ -38,6 +60,17 @@ public class CustomStringConfiguration {
         @SuppressWarnings("unused")
         public Builder setMainVerbStringResourceId(@StringRes final int mainVerbStringResId) {
             mainVerbStringResourceId = mainVerbStringResId;
+            return this;
+        }
+
+        /**
+         * Add the possibility to add a custom Title in payment vault screen.
+         *
+         * @param title Custom title to be setted
+         * @return builder to keep operating
+         */
+        public Builder setCustomPaymentVaultTitle(String title){
+            this.customPaymentVaultTitle = title;
             return this;
         }
